requirements.txt
absl-py==0.9.0
astroid==2.4.2
autopep8==1.5.4
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.4.1
cycler==0.10.0
Cython==0.29.17
-e git+https://github.com/facebookresearch/detectron2@04958b93e1232935e126c2fd9e6ccd3f57c3a8f3#egg=detectron2
future==0.18.2
fvcore @ git+https://github.com/facebookresearch/fvcore@ca3645612bb90b1b243a87d88dca671039d6f1b6
google-auth==1.14.1
google-auth-oauthlib==0.4.1
grpcio==1.28.1
idna==2.9
isort==5.6.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
matplotlib==3.2.1
mccabe==0.6.1
mock==4.0.2
numpy==1.13.3
oauthlib==3.1.0
Pillow==7.1.2
portalocker==1.7.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools @ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
pycodestyle==2.6.0
pydot==1.4.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
toml==0.10.2
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tqdm==4.45.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.7
